requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
bcrypt=4.0.1=pypi_0
ca-certificates=2023.08.22=hca03da5_0
cachetools=5.3.2=pypi_0
certifi=2023.7.22=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
cryptography=41.0.5=pypi_0
db-dtypes=1.1.1=pypi_0
google-api-core=2.14.0=pypi_0
google-auth=2.23.4=pypi_0
google-auth-oauthlib=1.1.0=pypi_0
google-cloud-bigquery=3.13.0=pypi_0
google-cloud-bigquery-storage=2.22.0=pypi_0
google-cloud-core=2.3.3=pypi_0
google-crc32c=1.5.0=pypi_0
google-resumable-media=2.6.0=pypi_0
googleapis-common-protos=1.61.0=pypi_0
grpcio=1.59.2=pypi_0
grpcio-status=1.59.2=pypi_0
idna=3.4=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_0
ncurses=6.4=h313beb8_0
numpy=1.24.4=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.12=h1a28f6b_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
pandas-gbq=0.19.2=pypi_0
paramiko=3.3.1=pypi_0
pip=23.3=py38hca03da5_0
proto-plus=1.22.3=pypi_0
protobuf=4.25.0=pypi_0
pyarrow=14.0.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydata-google-auth=1.8.2=pypi_0
pynacl=1.5.0=pypi_0
pysftp=0.2.9=pypi_0
python=3.8.18=hb885b13_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
pytz=2023.3.post1=pypi_0
readline=8.2=h1a28f6b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
setuptools=68.0.0=py38hca03da5_0
six=1.16.0=pypi_0
sqlite=3.41.2=h80987f9_0
tk=8.6.12=hb8d0fd4_0
tzdata=2023.3=pypi_0
urllib3=2.1.0=pypi_0
wheel=0.41.2=py38hca03da5_0
xz=5.4.2=h80987f9_0
zlib=1.2.13=h5a0b063_0
